+++ amfiteatr_rl/src/policy/actor_critic/discrete.rs
@@ -210,7 +210,7 @@ ContextDecodeIndexI64<ActionEncoding> + ContextEncodeIndexI64<ActionEncoding>{
         Ok((log_prob, entropy, a2c_net.critic))
 
          */
-        let critic_actor= self.network.net()(info_set_batch);
+        let critic_actor= self.network.operator()(self.network.var_store(), info_set_batch);
 
         let batch_logprob = critic_actor.batch_log_probability_of_action::<S>(
             action_param_batches,