+++ amfiteatr_rl/src/torch_net/network.rs
@@ -9,7 +9,8 @@ use std::default::Default;
 /// This structure shortens some steps of setting and provides some helpful functions - especially
 /// [`build_optimiser`](NeuralNet::build_optimizer).
 pub struct NeuralNet<Output: NetOutput>{
-    net: Box<dyn Fn(&Tensor) -> Output + Send >,
+    //net: Box<dyn Fn(&Tensor) -> Output + Send >,
+    operator: Box<dyn Fn(&VarStore, &Tensor) -> Output + Send>,
     var_store: VarStore,
     //_input: PhantomData<Input>,
 }
@@ -45,25 +46,26 @@ pub type MultiDiscreteNet = NeuralNet<MultiDiscreteTensor>;
 /// let device = Device::cuda_if_available();
 /// let var_store = VarStore::new(device);
 /// let number_of_actions = 33_i64;
-/// let neural_net = A2CNet::new(var_store, |path|{
-///     let seq = nn::seq()
-///         .add(nn::linear(path / "input", 16, 128, Default::default()))
-///         .add(nn::linear(path / "hidden", 128, 128, Default::default()));
-///     let actor = nn::linear(path / "al", 128, number_of_actions, Default::default());
-///     let critic = nn::linear(path / "cl", 128, 1, Default::default());
-///     let device = path.device();
-///     {move |xs: &Tensor|{
-///         let xs = xs.to_device(device).apply(&seq);
+/// let neural_net = A2CNet::new(var_store, Box::new(
+///     move |vs: &VarStore, tensor: &Tensor| {
+///         let seq = nn::seq()
+///         .add(nn::linear(vs.root() / "input", 16, 128, Default::default()))
+///         .add(nn::linear(vs.root() / "hidden", 128, 128, Default::default()));
+///     let actor = nn::linear(vs.root() / "al", 128, number_of_actions, Default::default());
+///     let critic = nn::linear(vs.root() / "cl", 128, 1, Default::default());
+///     let device = vs.device();
+///     let xs = tensor.to_device(device).apply(&seq);
 ///         //(xs.apply(&critic), xs.apply(&actor))
-///         TensorActorCritic{critic: xs.apply(&critic), actor: xs.apply(&actor)}
-///     }}
+///     TensorActorCritic{critic: xs.apply(&critic), actor: xs.apply(&actor)}
 ///
-/// });
+///     }
+/// ));
 ///
 /// let optimizer = neural_net.build_optimizer(Adam::default(), 0.01);
 /// ```
 impl<Output: NetOutput> NeuralNet< Output>{
 
+    /*
     pub fn new<
         N: 'static + Send + Fn(&Tensor) -> Output,
         F: Fn(&Path) -> N>
@@ -78,9 +80,18 @@ impl<Output: NetOutput> NeuralNet< Output>{
         }
     }
 
+     */
+
+    pub fn new(var_store: VarStore, operator: Box<dyn Fn(&VarStore, &Tensor) -> Output + Send>) -> Self{
+        NeuralNet { operator, var_store }
+    }
+
+    /*
     pub fn new_from_box(var_store: VarStore, network_f: Box<dyn Fn(&Tensor) -> Output + Send>) -> Self{
         Self{var_store, net: network_f}
     }
+
+     */
 
 
     /// Build optimiser for network, given `OptimizerConfig`. Uses [`VarStore`] stored in [`NeuralNet`] struct;
@@ -97,17 +108,23 @@ impl<Output: NetOutput> NeuralNet< Output>{
     /// use amfiteatr_rl::torch_net::NeuralNet;
     /// let device = Device::cuda_if_available();
     /// let var_store = VarStore::new(device);
-    /// let neural_net = NeuralNet::new(var_store, |path|{
+    /// let neural_net = NeuralNet::new(var_store, Box::new(|vs: &VarStore, tensor: &Tensor|{
     ///     let seq = nn::seq()
-    ///         .add(nn::linear(path / "input", 32, 4, Default::default()));
-    ///     move |tensor|{tensor.apply(&seq)}
+    ///         .add(nn::linear(vs.root() / "input", 32, 4, Default::default()));
+    ///     tensor.apply(&seq)
     ///
-    /// });
+    /// }));
     /// let input_tensor = Tensor::zeros(32, (Kind::Float, device));
-    /// let output_tensor = (neural_net.net())(&input_tensor);
+    /// let output_tensor = (neural_net.operator())(neural_net.var_store(),  &input_tensor);
     /// assert_eq!(output_tensor.size(), vec![4]);
     /// ```
-    pub fn net(&self) -> &(dyn Fn(&Tensor) -> Output + Send){&self.net}
+    pub fn operator(&self) -> &(dyn Fn(&VarStore, &Tensor) -> Output + Send){&self.operator}
+
+    /*pub fn net(&self) -> Box<dyn Fn(&Tensor) -> Output + Send>{
+        Box::new(|tensor| {
+            (self.operator())(&self.var_store, tensor)
+        })
+    }*/
 
     pub fn device(&self) -> Device{
         self.var_store.device()
@@ -139,6 +156,9 @@ pub struct NeuralNetTemplate<
 
 
 
+#[deprecated(since = "0.13.0", note = "Please build `NeuralNetTemplate` with function `new` using operator Box.\
+The whole structure does not make sence anymore since NuralNet use `Box<dyn Fn(&VarStore, &Tensor) -> Output>` \
+and not `Box<dyn Fn(&Tensor) -> Output + Send >` which led to problems with constructing dynamic network shape without templating them in closures.")]
 impl<
     O: NetOutput,
     N: 'static + Send + Fn(&Tensor) -> O,
@@ -153,16 +173,7 @@ NeuralNetTemplate<O, N, F>{
             net_closure
         }
     }
-    /*
-    pub fn new_actor_critic<H: Fn(Sequential) -> Sequential> -> Self(
-        input_dim: i64, actor_dim: i64, critic_dim: i64, hidden: H){
 
-        NeuralNetTemplate::new(|path|{
-            let seq = nn::seq()
-                .add(nn::linear)
-        })
-
-    }*/
 
     /// When you  do something like this it will fail to compile
     /// ```compile_fail
@@ -186,21 +197,27 @@ NeuralNetTemplate<O, N, F>{
     /// use tch::{Device, nn};
     /// use tch::nn::VarStore;
     /// use amfiteatr_rl::torch_net::{NeuralNet, NeuralNetTemplate};
+    ///
     /// let nc = NeuralNetTemplate::new(|path|{
     ///     let seq = nn::seq()
     ///         .add(nn::linear(path / "input", 32, 4, Default::default()));
     ///     move |tensor| {tensor.apply(&seq)}
     /// });
-    /// let (vs1, vs2) = (VarStore::new(Device::Cpu), VarStore::new(Device::Cpu));
     /// let closure = nc.get_net_closure();
-    /// let net1 = NeuralNet::new(vs1, closure);
-    /// let net2 = NeuralNet::new(vs2, closure);
+    ///
+    ///
+    /// let (vs1, vs2) = (VarStore::new(Device::Cpu), VarStore::new(Device::Cpu));
+    ///
+    /// // let net1 = NeuralNet::new(vs1, closure);
+    /// // let net2 = NeuralNet::new(vs2, closure);
     ///
     ///
     /// ```
     pub fn get_net_closure(&self) -> F{
         self.net_closure.clone()
     }
+
+
 }
 
 
@@ -287,6 +304,70 @@ impl NeuralNetActorCritic{
 }
 
 */
+
+
+pub fn build_network_operator_ac(layers: Vec<Layer>, input_shape:  Vec<i64>, actor_shape: i64)
+                                 ->  Box<dyn Fn(&VarStore, &Tensor) -> TensorActorCritic + Send>
+{
+
+
+
+    if !layers.is_empty() {
+        Box::new(move |vs, tensor|{
+            let mut seq = nn::seq();
+            let mut current_dim = input_shape[0];
+            let mut next_dim = input_shape[0];
+            let mut last_dim = None;
+            let mut layer = layers[0].clone();
+            last_dim = Some(layer.clone());
+            seq = match layer {
+                Layer::Relu => { seq.add_fn(|x| x.relu()) },
+                Layer::Tanh => { seq.add_fn(|x| x.tanh()) },
+                Layer::Linear(output) => {
+                    next_dim = output;
+                    seq.add(
+                        nn::linear(vs.root() / "0", current_dim, output, Default::default())
+                    )
+                }
+            };
+            current_dim = next_dim;
+            for (i, new_layer) in layers.iter().enumerate().skip(1) {
+                seq = match new_layer {
+                    Layer::Relu => { seq.add_fn(|x| x.relu()) },
+                    Layer::Tanh => { seq.add_fn(|x| x.tanh()) },
+                    Layer::Linear(output) => {
+                        next_dim = *output;
+                        seq.add(
+                            nn::linear(vs.root() / &format!("lin_{}", i), current_dim, *output, Default::default()))
+                    }
+                };
+                current_dim = next_dim;
+            }
+            let (actor, critic) = (
+                nn::linear(vs.root() / "actor", current_dim, actor_shape, Default::default()),
+                nn::linear(vs.root() / "critic", current_dim, 1, Default::default())
+            );
+
+
+
+            let device = vs.device();
+            let xs = tensor.to_device(device).apply(&seq);
+            TensorActorCritic {
+                critic: xs.apply(&critic),
+                actor: xs.apply(&actor),
+            }
+
+
+        })
+
+
+
+    } else{
+        panic!("Empty layer sequence")
+    }
+
+}
+
 /*
 pub fn create_net_model_discrete_ac<'a>(vs: &'a VarStore, layers: &[Layer], input_shape: &[i64], actor_shape: i64)
     ->  Box<dyn Fn(&Tensor) -> TensorActorCritic + Send + 'a>